+++ lib/src/map/built_map.dart
@@ -47,7 +47,7 @@ class BuiltMap<K, V> {
   /// Converts to a [MapBuilder] for modification.
   ///
   /// The `BuiltMap` remains immutable and can continue to be used.
-  MapBuilder toBuilder() => new MapBuilder<K, V>(this);
+  MapBuilder<K, V> toBuilder() => new MapBuilder<K, V>(this);
 
   /// Converts to a [MapBuilder], applies updates to it, and builds.
   BuiltMap<K, V> rebuild(updates(MapBuilder<K, V> builder)) =>