+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changelog
 
-## Unreleased
+## 1.0.0
 
 - Fix missing generics on some return types.
 - Fix BuiltList and BuiltSet "contains" method, should take Object, not E.
 - Add removeAll and retainAll methods to SetBuilder.
 - Add BuiltList.toBuiltSet() and BuiltSet.toBuiltList().
+- Add addIterable methods to map builders.
 
 ## 0.4.0